+++ site/components/Renderer.jsx
@@ -1,9 +1,11 @@
 import React from "react";
-import { Button, Panel, Table } from "react-bootstrap";
-import { Maps } from "robe-react-commons";
+import {Button, ButtonGroup, Panel, Table} from "react-bootstrap";
+import {Maps} from "robe-react-commons";
 import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
 import Highlight from "react-highlight";
 import Progress from "progress/Progress";
+import FaIcon from "faicon/FaIcon";
+import Toast from "toast/Toast";
 
 
 export default class Renderer extends ShallowComponent {
@@ -43,6 +45,15 @@ export default class Renderer extends ShallowComponent {
     static defaultProps = {
         json: {}
     };
+
+    static clipboardStyle = {
+        position: "absolute",
+        display: "inline-table",
+        marginTop: 5,
+        opacity: 0.6,
+        right: 36
+    };
+
     /* eslint no-useless-constructor: 0*/
     constructor(props) {
         super(props);
@@ -51,18 +62,43 @@ export default class Renderer extends ShallowComponent {
         };
     }
 
-    render(): Object {
+    render():Object {
         let highlight = undefined;
         if (this.state.showCode) {
-            highlight = (<Highlight className="javascript">
-                {this.props.code}
-            </Highlight>);
+            highlight = (
+                <div>
+                    <div className="pull-right">
+                        <ButtonGroup
+                            style={Renderer.clipboardStyle}>
+                            <Button
+                                bsSize="xsmall"
+                                onClick={this.__copyToClipboard}>
+                                <FaIcon code="fa-clipboard"/>
+                            </Button>
+                            <Button
+                                bsSize="xsmall"
+                                onClick={this.__copyToClipboard}>
+                                <FaIcon code="fa-file-text-o"/>
+                            </Button>
+                            <Button
+                                bsSize="xsmall"
+                                onClick={this.__copyToClipboard}>
+                                <FaIcon code="fa-download"/>
+                            </Button>
+                        </ButtonGroup>
+                    </div>
+                    <Highlight className="javascript">
+                        {this.props.code}
+                    </Highlight>
+                </div>
+            );
         }
 
         let codeSection = this.props.code ?
             (<div>
                 {highlight}
-                <Button bsStyle="link" bsSize="xsmall" className="pull-right" onClick={this.__toogleCode}>{(this.state.showCode ? "Hide" : "Show") + " Code"}</Button>
+                <Button bsStyle="link" bsSize="xsmall" className="pull-right"
+                        onClick={this.__toogleCode}>{(this.state.showCode ? "Hide" : "Show") + " Code"}</Button>
             </div>) : undefined;
         return (
             <div>
@@ -80,20 +116,41 @@ export default class Renderer extends ShallowComponent {
             </div >);
     }
 
-    __toogleCode = () => {
+    __toogleCode() {
         this.setState({
             showCode: !this.state.showCode
         });
     }
 
-    __renderPropsTable(data: Object): Array {
+    __copyToClipboard() {
+        let textField = document.createElement('textarea');
+        textField.innerHTML = this.props.code;
+
+        document.body.appendChild(textField);
+
+        let range = document.createRange();
+        range.selectNode(textField);
+        textField.select();
+
+        window.getSelection().removeAllRanges();
+        window.getSelection().addRange(range);
+        try {
+            document.execCommand('copy');
+            Toast.success("Copy successful.")
+        } catch (err) {
+            console.log("Oops, unable to copy");
+        }
+        document.body.removeChild(textField);
+    }
+
+    __renderPropsTable(data:Object):Array {
         if (data === undefined) {
             return undefined;
         }
 
         let rows = [];
 
-        Maps.forEach(data, (value: any, key: string) => {
+        Maps.forEach(data, (value:any, key:string) => {
             let type = value.type !== undefined ? value.type.name : "";
             let defaultVal = value.defaultValue !== undefined ? value.defaultValue.value : "";
             rows.push(<tr key={key}>
@@ -108,21 +165,22 @@ export default class Renderer extends ShallowComponent {
         return (
             <Table responsive striped bordered condensed>
                 <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Type</th>
-                        <th>Default</th>
-                        <th>Required</th>
-                        <th>Description</th>
-                    </tr>
+                <tr>
+                    <th>Name</th>
+                    <th>Type</th>
+                    <th>Default</th>
+                    <th>Required</th>
+                    <th>Description</th>
+                </tr>
                 </thead>
                 <tbody>
-                    {rows}
+                {rows}
                 </tbody>
             </Table>
         );
     }
-    __renderMethodsTable(data: Object): Array {
+
+    __renderMethodsTable(data:Object):Array {
         if (data === undefined) {
             return undefined;
         }
@@ -143,18 +201,19 @@ export default class Renderer extends ShallowComponent {
         return (
             <Table responsive striped bordered condensed>
                 <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Returns</th>
-                        <th>Description</th>
-                    </tr>
+                <tr>
+                    <th>Name</th>
+                    <th>Returns</th>
+                    <th>Description</th>
+                </tr>
                 </thead>
                 <tbody>
-                    {rows}
+                {rows}
                 </tbody>
             </Table>
         );
     }
+
     componentDidUpdate() {
         Progress.done();
     }