+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * robe-react-commons updated
 * Added Copy Button to Highlight
 * update style in SelectInput
+* js-criteria and robe-react-common version incremented.
+
 
 ## 1.1.7
 * robe-react-commons updated